fix(scanArea): migrate filters outside the layer
