+++ src/features/scanArea/index.js
@@ -1,3 +1,4 @@
 // @ts-check
 
+export * from './LegacyMigration'
 export * from './ScanAreaTile'
