+++ src/pages/map/components/Container.jsx
@@ -6,6 +6,7 @@ import { useMemory } from '@store/useMemory'
 import { useStorage } from '@store/useStorage'
 import { useMapStore } from '@store/useMapStore'
 import { ScanOnDemand } from '@features/scanner'
+import { LegacyScanAreaMigration } from '@features/scanArea'
 import { WebhookMarker, WebhookAreaSelection } from '@features/webhooks'
 import { ActiveWeather } from '@features/weather'
 import { timeCheck } from '@utils/timeCheck'
@@ -57,6 +58,7 @@ export function Container() {
       preferCanvas
     >
       <Effects />
+      <LegacyScanAreaMigration />
       <ControlledTileLayer />
       <ControlledZoomLayer />
       <ControlledLocate />
